refactor: move validateview function
