+++ cause/views.py
@@ -58,14 +58,3 @@ class CausesPatch(ViewSet):
         response_serializer = CausesResponse(cause)
 
         return Response(response_serializer.data)
-
-@permission_classes([])
-class ValidateView(APIView):
-    @extend_schema(
-        description='Run Root Cause Analysis for a specific question and row',
-        responses=CausesResponse,
-    )
-    def patch(self, request, question_id):
-        updated_causes = CausesService.validate(self=CausesService, question_id=question_id)
-        serializer = CausesResponse(updated_causes, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
